Split topic grant tasks
